requirements.txt
replicate
python-dotenv
requests
google-generativeai
